Move `OrganisationsPresenter` into `admin` folder to reflect usage
